Deutlich einfacherer Vorschlag für Transaktionen

+++ src/main/java/com/WWI16AMA/backend_api/Events/ExtTransactionEvent.java
@@ -5,16 +5,16 @@ import com.WWI16AMA.backend_api.Member.Member;
 
 public class ExtTransactionEvent extends MemberEvent {
 
-    private int amount;
+    private double amount;
     private Transaction.FeeType type;
 
-    public ExtTransactionEvent(Member member, int amount, Transaction.FeeType type) {
+    public ExtTransactionEvent(Member member, double amount, Transaction.FeeType type) {
         super(member);
         this.amount = amount;
         this.type = type;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
